functions/menu_image/requirements.txt
appwrite==11.1.0
google-genai==1.30.0
